feat: Add OptimizationInterface for backend optimization management
- Introduced OptimizationInterface class to serve as an abstraction layer for interacting with EOS and EVCC Opt optimization backends.
- Implemented methods for optimization, control data examination, and scheduling management.
- Added functionality to calculate the next optimal run time and retrieve EOS version from the backend.
- Created unit tests for OptimizationInterface, covering backend selection, response handling, and error management.
- Included a dummy backend for testing integration without actual backend dependencies.
